Major utils restructure
