+++ src/utils/colorUtils.ts
@@ -8,7 +8,38 @@ export const baseColors = {
   primaryGrey: '#E0E0E0',
 }
 
+/**
+ * Create a list of colors based on the number of objects
+ * Rainbow style for now, we can limit this range later
+ * @param n number of colors
+ * @returns list of HSL colors
+ */
+export function generateColors(n: number): string[] {
+  const colors: string[] = []
+  for (let i = 0; i < n; i++) {
+    const hue = Math.round(360 * i / n)
+    colors.push(`hsl(${hue}, 100%, 80%)`)
+  }
+  return colors
+}
 
+/**
+ * Convert hsl to hex, https://stackoverflow.com/questions/36721830/convert-hsl-to-rgb-and-hex
+ * @param h hue
+ * @param s saturation
+ * @param l lightness
+ * @returns Hex color
+ */
+export function hslToHex(h: number, s: number, l: number) {
+  l /= 100
+  const a = s * Math.min(l, 1 - l) / 100
+  const f = n => {
+    const k = (n + h / 30) % 12
+    const color = l - a * Math.max(Math.min(k - 3, 9 - k, 1), -1)
+    return Math.round(255 * color).toString(16).padStart(2, '0')   // convert to Hex and prefix "0" if needed
+  }
+  return `#${f(0)}${f(8)}${f(4)}`
+}
 
 /**
  * Get gradient color scale for custom color length
